:recycle:	refactor: Create app
